added checking of database field names against client record field names
Seemed like it would be nice to do a run-time check that the
field names in the database were as expected. So it was
done. (The number of fields was already being checked.)

Assuming neither the database nor the client allows
duplicate field names in a single record, this change
future-proofs the system. In case either the client or the
database changes alone, a run-time error is guaranteed.

The current implementation is a bit ugly, but I think we can
live with it.